+++ Xappy/Xappy/Content/Scenarios/OtherLogin/Controls/GradientButton.xaml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ContentView xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Xappy.Content.Scenarios.Login2.Controls"
-             x:Class="Xappy.Content.Scenarios.Login2.Controls.GradientButton"
+             xmlns:controls="clr-namespace:Xappy.Content.Scenarios.OtherLogin.Controls"
+             x:Class="Xappy.Content.Scenarios.OtherLogin.Controls.GradientButton"
              x:Name="This">
     <controls:PressableView Pressed="Handle_Pressed"
                             Released="Handle_Released">
