+++ Xappy/Xappy/Content/Scenarios/OtherLogin/Controls/GradientView.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Xamarin.Forms;
 
-namespace Xappy.Content.Scenarios.Login2.Controls
+namespace Xappy.Content.Scenarios.OtherLogin.Controls
 {
     public enum GradientDirection
     {
