+++ Xappy/Xappy/Content/Scenarios/OtherLogin/Controls/PressableView.cs
@@ -2,7 +2,7 @@
 
 using Xamarin.Forms;
 
-namespace Xappy.Content.Scenarios.Login2.Controls
+namespace Xappy.Content.Scenarios.OtherLogin.Controls
 {
     public class PressableView : ContentView
     {
