+++ Xappy/Xappy/Content/Scenarios/OtherLogin/LoginPage.xaml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Xappy.Content.Scenarios.Login2.Controls"
-             x:Class="Xappy.Content.Scenarios.Login2.LoginPage"
+             xmlns:controls="clr-namespace:Xappy.Content.Scenarios.OtherLogin.Controls"
+             x:Class="Xappy.Content.Scenarios.OtherLogin.LoginPage"
              xmlns:lottie="clr-namespace:Lottie.Forms;assembly=Lottie.Forms"
              Title="Login"
              Shell.TitleColor="White"
