+++ Xappy/Xappy/Xappy.csproj
@@ -37,8 +37,9 @@
     <Folder Include="Content\Scenarios\ToDo\" />
     <Folder Include="Content\About\" />
     <Folder Include="Content\Common\" />
-    <Folder Include="Content\Scenarios\Login2\" />
-    <Folder Include="Content\Scenarios\Login2\Controls\" />
+    <Folder Include="Content\Scenarios\OtherLogin\" />
+    <Folder Include="Content\Scenarios\OtherLogin\Controls\" />
+    <Folder Include="Content\Scenarios\Login\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="ApiConstants.cs" />
